chore: move InProcNet to testutil
